backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
Pillow==9.0.0
python-dotenv==0.21.0
reportlab==3.6.12
psycopg2-binary==2.8.6
gunicorn==20.0.4